+++ examples/src/Server/Mythology/Character/Deity.hs
@@ -1,7 +1,9 @@
+{-# LANGUAGE DeriveAnyClass    #-}
 {-# LANGUAGE DeriveGeneric     #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeFamilies      #-}
 {-# LANGUAGE TypeOperators     #-}
+
 
 module Server.Mythology.Character.Deity
   ( Deity(..)
@@ -9,7 +11,6 @@ module Server.Mythology.Character.Deity
   )
 where
 
-import           Data.Morpheus.Kind             ( OBJECT )
 import           Data.Morpheus.Types            ( GQLType(..) )
 import           Data.Text                      ( Text )
 import           GHC.Generics                   ( Generic )
@@ -19,11 +20,7 @@ data Deity = Deity
   { fullName :: Text -- Non-Nullable Field
   , power    :: Maybe Text -- Nullable Field
   , realm    :: Realm
-  } deriving (Generic)
-
-instance GQLType Deity where
-  type KIND Deity = OBJECT
-  description _ = Just "Custom Description for Client Defined User Type"
+  } deriving (Generic,GQLType)
 
 dbDeity :: Text -> Maybe Text -> IO (Either String Deity)
 dbDeity _ _ = return $ Right $ Deity { fullName = "Morpheus"