+++ examples/src/Server/Mythology/Character/Human.hs
@@ -1,13 +1,13 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE TypeFamilies  #-}
-{-# LANGUAGE TypeOperators #-}
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric  #-}
+{-# LANGUAGE TypeFamilies   #-}
+{-# LANGUAGE TypeOperators  #-}
 
 module Server.Mythology.Character.Human
   ( Human(..)
   )
 where
 
-import           Data.Morpheus.Kind             ( OBJECT )
 import           Data.Morpheus.Types            ( GQLType(..) )
 import           Data.Text                      ( Text )
 import           GHC.Generics                   ( Generic )
@@ -15,8 +15,5 @@ import           Server.Mythology.Place.Places  ( City(..) )
 
 data Human = Human
   { name :: Text
-  , home :: City
-  } deriving (Generic)
-
-instance GQLType Human where
-  type KIND Human = OBJECT
+  , bornAt :: City
+  } deriving (Generic,GQLType)