+++ changelog.md
@@ -1,3 +1,131 @@
+## [0.7.2] -
+
+### Added
+
+- Automatic Type Inference (only for Object, Union and Enum)
+
+```hs
+
+data Realm  =
+    Sky
+  | Sea
+  | Underworld
+    deriving (Generic, GQLType)
+
+data Deity  = Deity{
+    fullName:: Text,
+    realm:: Realm
+  } deriving (Generic, GQLType)
+
+data Character  =
+    CharacterDeity Deity -- Only <tyconName><conName> should generate direct link
+  -- RECORDS
+  | Creature { creatureName :: Text, creatureAge :: Int }
+  --- Types
+  | SomeDeity Deity
+  | CharacterInt Int
+  | SomeMutli Int Text
+  --- ENUMS
+  | Zeus
+  | Cronus deriving (Generic, GQLType)
+
+
+```
+
+will generate schema:
+
+```gql
+enum Realm {
+  Sky
+  Sea
+  Underworld
+}
+
+type Deity {
+  fullName: String!
+  realm: Realm!
+}
+
+union Character =
+    Deity
+  | Creature
+  | SomeDeity
+  | CharacterInt
+  | SomeMutli
+  | CharacterEnumObject
+
+type Creature {
+  creatureName: String!
+  creatureAge: Int!
+}
+
+type SomeDeity {
+  _0: Deity!
+}
+
+type CharacterInt {
+  _0: Int!
+}
+
+type SomeMutli {
+  _0: Int!
+  _1: String!
+}
+
+# enum
+type CharacterEnumObject {
+  enum: CharacterEnum!
+}
+
+enum CharacterEnum {
+  Zeus
+  Cronus
+}
+```
+
+rules:
+
+- haskell union type with only empty constructors (e.g `Realm`), will generate graphql `enum`
+- haskell record without union (e.g `Deity`), will generate graphql `object`
+- namespaced Unions: `CharacterDeity` where `Character` is TypeConstructor and `Deity` referenced object (not scalar) type: will be generate regular graphql Union
+
+  ```gql
+  union Character =
+        Deity
+      | ...
+  ```
+
+- for union recrods (`Creature { creatureName :: Text, creatureAge :: Int }`) will be referenced in union type, plus type `Creature`will be added in schema.
+
+  e.g
+
+  ```gql
+    union Character =
+      ...
+      | Creature
+      | ...
+
+    type Creature {
+      creatureName : String!
+      creatureAge: Int!
+    }
+
+  ```
+
+  - all empty constructors in union will be summed in type `<tyConName>Enum` (e.g `CharacterEnum`), this enum will be wrapped in `CharacterEnumObject` and this type will be added to union `Character`. as in example above
+
+  - there is only types left with form `TypeName Type1 2Type ..`(e.g `SomeDeity Deity` ,`CharacterInt Int`, `SomeMutli Int Text`),
+
+    morpheus will generate objet type from it:
+
+    ```gql
+    type TypeName {
+      _0: Type1!
+      _1: Type2!
+      ...
+    }
+    ```
+
 ## [0.7.1] - 26.11.2019
 
 - max bound icludes: support-megaparsec-8.0